requirements.txt
openai
langchain
python-dotenv
gitpython
deeplake
tiktoken